+++ lib/logstash/inputs/signalsciences.rb
@@ -25,12 +25,9 @@ class LogStash::Inputs::Signalsciences < LogStash::Inputs::Base
   config :site, :validate => :string, :default => "not_provided"
   # Number of seconds in the past to filter data on
   # This value will also be used to set the interval at which the API is polled.
-  config :from, :validate => :number, :default => 600
+  config :from, :validate => :number, :default => 300
   # Debug for plugin development.
   config :debug, :validate => :boolean, :default => false
-
-  # Set how frequently messages should be sent.
-  #
   # The default, `600`, means fetch data every 10 minutes.
   config :interval, :validate => :number, :default => 600
 
@@ -43,10 +40,10 @@ class LogStash::Inputs::Signalsciences < LogStash::Inputs::Base
     @get = Net::HTTP::Get.new("/api/v0/corps/#{@corp}/sites/#{@site}/feed/requests")
     @http.use_ssl = true
 
-    # check if from value is 5 minutes or less
-    if @from <= 300
-      @logger.warn("from value is 5 minutes or less, increasing from value to 10 minutes.")
-      @from = 600
+    # check if from value is less than 1 min
+    if @from < 60
+      @logger.warn("from value is less than 1 min, increasing from value to 1 minute.")
+      @from = 60
     end
 
     # check if from value is greater than 24 hours
@@ -55,10 +52,10 @@ class LogStash::Inputs::Signalsciences < LogStash::Inputs::Base
       @from = 86400
     end
 
-    # set interval to value of from @from minus one minute
-    interval = @from
+    # set interval to value of from @from minus five minutes
+    @interval = @from
 
-    @logger.info("Fetching Signal Sciences request data every #{interval / 60} minutes.")
+    @logger.info("Fetching Signal Sciences request data every #{@interval / 60} minutes.")
   end # def register
 
   def run(queue)
@@ -68,6 +65,8 @@ class LogStash::Inputs::Signalsciences < LogStash::Inputs::Base
 
       if fetch(queue)
         @logger.info("Requests feed retreived successfully.")
+      else
+        @logger.warn("Problem retreiving request!")
       end
 
       # because the sleep interval can be big, when shutdown happens
@@ -75,6 +74,7 @@ class LogStash::Inputs::Signalsciences < LogStash::Inputs::Base
       # Stud.stoppable_sleep will frequently evaluate the given block
       # and abort the sleep(@interval) if the return value is true
       #Stud.stoppable_sleep(@interval) { stop? }
+      @logger.info("Sleep #{@interval}")
       Stud.stoppable_sleep(@interval) { stop? }
     end # loop
   end # def run
@@ -82,6 +82,7 @@ class LogStash::Inputs::Signalsciences < LogStash::Inputs::Base
   def fetch(queue)
     begin
       response = @http.request(@login)
+      @logger.warn("login response: #{response.code}")
     rescue
       @logger.warn("Could not reach API endpoint to login!")
       return false
@@ -98,12 +99,19 @@ class LogStash::Inputs::Signalsciences < LogStash::Inputs::Base
       token = json['token']
       # Both the from and until parameters must fall on full minute boundaries,
       # see https://docs.signalsciences.net/faq/extract-your-data/.
-      t = Time.now.strftime("%Y-%m-%d %H:%M:0")
+      t = Time.now.utc.strftime("%Y-%m-%d %H:%M:0")
       dt = DateTime.parse(t)
       timestamp_until = dt.to_time.to_i - 300 # now - 5 minutes
-      timestamp_from = (timestamp_until - @from) - 300 # @from - 5 minutes
+      timestamp_from = (timestamp_until - @from) # @until - @from
+
+      if @debug
+        hfrom = Time.at(timestamp_from).to_datetime
+        huntil = Time.at(timestamp_until).to_datetime
+        @logger.info("From #{hfrom} Until #{huntil}")
+      end
 
       # Set up iniital get request and initial next_uri
+      @logger.info("Requesting data: /api/v0/corps/#{@corp}/sites/#{@site}/feed/requests?from=#{timestamp_from}&until=#{timestamp_until}")
       get = Net::HTTP::Get.new("/api/v0/corps/#{@corp}/sites/#{@site}/feed/requests?from=#{timestamp_from}&until=#{timestamp_until}")
       next_uri = "not empty on first pass"
 
@@ -122,7 +130,7 @@ class LogStash::Inputs::Signalsciences < LogStash::Inputs::Base
         #check for message, error, e.g. missing query string parameter
         if json.has_key? "message"
           # some error occured, report it.
-          @logger.warn("get: #{json['message']} #{token}")
+          @logger.warn("Error accessing API (#{token}), status code: #{response.code} with message: #{json['message']}")
           return false
 
         else
@@ -135,7 +143,9 @@ class LogStash::Inputs::Signalsciences < LogStash::Inputs::Base
               payload['headersIn'].each { |k,v| temp[k] = v }
               payload["headerIn"] = temp
             rescue NoMethodError
-              @logger.info("payload['headersIn'] is empty for id #{payload['id']}, skipping append.")
+              if @debug
+                @logger.debug("payload['headersIn'] is empty for id #{payload['id']}, skipping append.")
+              end
             end
 
             # explode headersOut out to headerOut entries
@@ -144,7 +154,9 @@ class LogStash::Inputs::Signalsciences < LogStash::Inputs::Base
               payload['headersOut'].each { |k,v| temp[k] = v }
               payload["headerOut"] = temp
             rescue NoMethodError
-              @logger.info("payload['headersOut'] is empty for id #{payload['id']}, skipping append.")
+              if @debug
+                @logger.info("payload['headersOut'] is empty for id #{payload['id']}, skipping append.")
+              end
             end
 
             # explode tags out to tag entries
@@ -155,6 +167,7 @@ class LogStash::Inputs::Signalsciences < LogStash::Inputs::Base
             payload['tag'] = temp
 
             # add the event
+            
             event = LogStash::Event.new("message" => payload, "host" => @host)
 
             decorate(event)
@@ -163,6 +176,9 @@ class LogStash::Inputs::Signalsciences < LogStash::Inputs::Base
 
           # get the next uri value
           next_uri = json['next']['uri']
+          if @debug
+            logger.info("Next URI: #{next_uri}")
+          end
 
           # continue retreiving next_uri if it exists
           if !next_uri.empty?