+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0
+
+- Changed time to UTC
+- Changed default from value to 5 min
+- Updated from value to allow for pulling data every 1 min
+- Added additional logging for debug
+
 ## 0.1.1
 
 - Fixed references to vulnerable dependencieas in Gemfile.lock